Try to fix a reentrancy bug in chess-display-move.
The problem occurs with engine modules which reply to move events with move
events directly, i.e. cvhess-ai and chess-uci when replying with a book move.
Since chess-display-move first performs chess-game-move followed by
chess-display-paint-move, we end up with the plies getting painted in reverse.
This is visible when pieces are taken.
If we paint the move first and then emit the move event on the game,
the problem goes away.