eglot-completion-at-point: Follow specification more closely
This is a WIP.

It seems eglot assumed that the triggerChacters couldn't be inside the
completion bounds.  This assumption does not hold with the
vscode-json-languageserver.  So the patched version of
eglot-completion-at-point asks the server for possible completions and
calculates the bounds from them (if the completions are consistent in
this regard.)

Additionally, vscode-json-languageserver returns completions with
different filterText and label fields.  The LSP specification is
usually quite vague, but not this time:

   filterText: A string that should be used when filtering a set of
   completion items. When `falsy` the label is used.

Although it never defines what it means by filtering.  Anyway, the new
test `json-basic` succeeds when the patch is applied, and fails
otherwise.  It assumes the server installed as root, e.g.:

   `sudo npm install -g vscode-json-languageserver`

Unfortunately, the code needs polishing.  eglot calls
textDocument/completion twice at the beginning of a completion.  But I
don't understand why it is necessary to fetch the completions from the
server more than once, so I haven't modified this part of the code.

I haven't run the rust and the eclipse tests, but hopefully those are
unaffected by this change.

Feedback on how to proceed is obviously welcome.  Thanks.