Fix #259: work around a bug in Emacs's change detection
The bug happens when using capitalize-word.  before-change-functions
will record e.g. the whole word's start and end, even though only the
first character has changed.  Not only is this longer than needed but
also inconsistent with what we get in after-change-functions, which
records just one-char-long change, so we must fix that.

* eglot.el (eglot--after-change): Fix end position sometimes.