WIP: Advertise the title of the current code action (#411)
If eglot-confirm-server-initiated-edits was nil and we had
"auto-confirming a single option", then users weren't informed about
the details of edits during eglot-code-actions.  This PR helps in this
regard.

The server's reply to a textDocument/codeAction may contain workspace
edits with titles, or commands with titles.  However, the client needs
to send the command the sever, and the server replies something, and
later it sends a workspace edit without a title.  Therefore, the new
variable `eglot--last-title' caches this information.