Per #474, #478: add more tests for `eglot--guess-contact'
Modify eglot--guessing-contact so that (1) each test is run once with
eglot--guessing-contact INTERACTIVE argument set to t, and again with
it set to nil; (2) executable-find is mocked, and (3) arguments passed
to read-shell-command are captured and made available to tests.

Co-authored-by: João Távora <joaotavora@gmail.com>

* eglot-tests.el (eglot--guessing-contact): Rework.
(eglot-server-programs-simple-existing-executable): New test.
(eglot-server-programs-simple-executable)
(eglot-server-programs-executable-multiple-major-modes)
(eglot-server-programs-executable-with-arg)
(eglot-server-programs-executable-with-args-and-autoport)
(eglot-server-programs-host-and-port)
(eglot-server-programs-host-and-port-and-tcp-args)
(eglot-server-programs-class-name-and-plist)
(eglot-server-programs-class-name-and-contact-spec)
(eglot-server-programs-function): Add new test assertions.