Try to make hover-multiline-doc-locus test pass on Travis
A batch emacs apparently runs with a frame-width of 10

   $ emacs -Q --batch --eval '(princ (frame-width))'
   10

This could possibly not be enough, for the hover-multiline-doc-locus
test, which uses pyls docstrings, to avoid truncation of
"datetime(...)" into "datetim...".  The failure doesn't happen
locally, but changing the assertion to check one character less isn't
very problematic.

* eglot-tests.el (hover-multiline-doc-locus): Try to make test
pass on Travis.