Propagate exec-path and process-environment when calling make-process
If either of these variables is set locally in a project or buffer,
e.g. using directory-local variables or packages like
`add-node-modules-path` and `envrc`, then the user will be able to
"see" and execute the language server program from their buffer, but
eglot will not be able to find it.

The appropriate fix is to explicitly propagate both vars to the
started process, so that it acts as if it were started from the
initiating buffer.