Fix eglot-completion-at-point so it can be used from the minibuffer
I have been playing with a `completing-read` version of
`completion-at-point`, and got it just about working, except with
`minibuffer-force-complete` was called I’d run into errors. Debugged it
to the `exit-function`, which was written assuming the current buffer
was still the code buffer. The fix was fairly trivial: change the
current buffer to the one from which the minibuffer was called, if
applicable. An alternative considered was to apply this logic in
`eglot--current-server-or-lose` and `eglot--TextDocumentPositionParams`,
which is were the errors came from, but that didn’t make sense.