call `eglot--maybe-activate-editing-mode` only on reachable buffers
The purpose of calling `eglot--maybe-activate-editing-mode` here is to
handle all buffers that could've been handled by the server we've just
set up *and* hadn't been doing so.

However, it wouldn't make sense to try to manage / edit a buffer that
is not reachable by the server. This happens when we're editing tramp
/ remote files, in which the server may be running in one
machine (say, my-remote-server) and a buffer is running in another
one (say, localhost)