Support optional :activeParameter property for SignatureInformation
* eglot.el (eglot--lsp-interface-alist): Add
SignatureInformation.activeParameter.

* eglot.el (eglot--sig-info): Prioritize
SignatureInformation.activeParameter over
SignatureHelp.activeParameter.

The SignatureInformation.activeParameter property was requested to
obviate the need for the UI to make extra requests when an alternate
signature was selected, since thanks to this feature, the signature
contains the info.  This seems to be particularly desirable in Python.

More info at https://github.com/microsoft/vscode/issues/94637.

SignatureInformation.activeParameter is new in version 3.16.0 of the
protocol.  When non-nil, it is used in place of
SignatureHelp.activeParameter.

See
https://microsoft.github.io/language-server-protocol/specifications/specification-current/#textDocument_signatureHelp