Uncover a test failure hidden by ert functino misuse
The test may be good, but it was running another instance of ert
inside it, meaning it would always pass, I think. It's better to have
both the normal and the TRAMP test call a function.  Then we can see
that the TRAMP automatic test is failing.

We have to understand why.