Add project-wide diagnostics
By adding all the diagnostics received from a server to a hash-table attached to
the server, we get the opportunity to access diagnostics for the whole workspace.

I have added one interesting entry point to this,
`eglot-find-workspace-diagnostics', that gives us the possibility to jump
directly to files with the associated diagnostics type.  This commit also
addresses the possibility to run other checkers through the newly introduced
hook.

* eglot.el (eglot-lsp-server): Add a new slot, `diagnostics', which is a
hash-table of all the diagnostics in a workspace attached to the current
server. KEY is the URI to the file in question and its associated unmodified
response from server.

* eglot.el (eglot--unreported-diagnostics): Remove this defvar-local, since
eglot now relies on diagnostics attached to the server.

* eglot.el (eglot--mode-line-format): Add information about project-wide
diagnostics to the mode line.  It follows the convention of flymake, with the
same propertization and fontification.

* eglot.el (eglot--update-diags): Utility function to update the server instance
with the payload received from the server.  Always overwrites everything, so
that the information stored in the instance is as fresh as possible.

* eglot.el (eglot--diagnostic-files-by-severity): Returns a list of files with
diagnostics of SEV kind.  It has no duplicates, but a file could have both
warnings and errors, and thus appear in to different invokations to this
function.

* eglot.el (eglot-find-workspace-diagnostics): The bread and butter of this
commit.  It adds `completing-read' to fetch files with diagnostics of the given
kind. Successive invokations to the universal argument yields different
diagnostic types.

* eglot.el (eglot-handle-notification): Move handling of the diagnostics to an
external hook, as well as simply storing the received diagnostics. Since we now
handle diagnostics for unvisited files, we can skip that debug logging.

* eglot.el (eglot-after-diagnostics-hook): New hook for updating of diagnostics
in the current buffer.

* eglot.el (eglot--update-diagnostics): New function to handle the received
payload of diagnostics to the current buffer.  Meant to run as a hook in
`eglot-after-diagnostics-hook'.

* eglot.el (eglot-flymake-backed): Run the new hook when this function is run
the first time. Other times, just run it directly from the hook handler through
`eglot-handle-notification'.
Add project-wide diagnostics