Support the which-func package, i.e. which-function-mode
which-func assumes that imenu is enabled and determines which function
your current point is in using imenu--index-alist.

However, it doesn’t understand special elements in imenu--index-alist
so those entries must be handled by a hook installed in
which-func-functions.

* eglot.el (eglot-which-func): Hook to determine the function which
your current point is in, based on Eglot’s special Imenu element.

* eglot.el (eglot--managed-mode): Add and remove the eglot-which-func
hook in which-func-functions.

* eglot-tests.el (basic-which-func): Test which-function-mode.