Correct the default-directory used for remote files
The return value of eglot--uri-to-path in remote projects used to be
the return value of file-remote-p concatenated with the (empty) file
URI, which strips the project prefix, and loses the location
information for dir-locals files.

Instead, now we use the entire project root - which at least returns
gets us as far as the project root dir (the file URI is still empty,
so deeper-nested dir-locals won't be used).