Add quirk support for initialize parameters
Certain language servers require non-standard parameters to be passed
as part of their initialization. Currently eglot does not support
providing additional base parameters outside of the
initializationOptions and capabilities.

Adding a generic function that defaults to an empty list allows users
to provide custom parameters to deal with such quirks.

An example is the CodeQL language server, which expects the
workspaceFolders parameter to set.