initialize workspaceFolders[0] to rootUri by default
Keeping `eglot-workspace-folders` as a generic still allows folks to
override this default behavior, but as per:

https://microsoft.github.io/language-server-protocol/specifications/specification-current/#initialize

both rootPath and rootUri are considered deprecated in favor of
workspaceFolders, I believe setting rootPath, rootUri, and
workspaceFolders[0] to rootUri, as part of the default initialize
parameters is in line with the current spec.

rootUri is set from `default-directory`, which is derived from
`project-root`. Future iterations of this could include filling this
list with `project-external-roots` as well for basic workspaceFolders
support as per @nemethf in:
https://github.com/joaotavora/eglot/pull/890#issuecomment-1073071461