Solve Flymake diags synch problems on "lazy" servers
A diagnostics-lazy server is one who doesn't re-report already
reported diagnostics when it received textDocument/didSave.

Such is the case of Clangd.  Before this change saving an
Eglot/Clang-managed buffer with some diagnostics caused the Flymake
indicator to display Wait[0 0] until some change was actually done to
the buffer.  That is because Flymake, by default wants diagnostics on
buffer save.

This change makes it so that when such a Flymake request comes in, it
always gets served immediately with the latest information.

Also it doesn't work to turn off flymake-start-on-save-buffer because
then if the LSP diagnostics do come in after the save (for some
reason), then Flymake won't have requested any diagnostics at all
, probaly because the modified bit is reset on save.

(eglot--maybe-activate-editing-mode)
(eglot-handle-notification publishDiagnostics): Don't overcook eglot--unreported-diagnostics.
(eglot-flymake-backend): Unconditionally immediately report something.
(eglot--report-to-flymake): Set eglot--last-reported-diagnostics.