Fix #941: Ensure exit-function of eglot-c-at-point runs on exact match
When the completion is exact match, exit-function should still run.

Say one is using auto-imports feature of pyright.  One types foo, and
triggers the completion.  There are two candidates: foo and foo_bar.  If
one chooses foo, the status would be 'exact' instead of 'finished', thus
exit-function is not executed, foo is not auto-imported.

* eglot.el (eglot-completion-at-point): Consider 'exact status.

Copyright-paperwork-exempt: Yes