For xreffed-to projects - adopt eglot server, if project doesn't have any
Eglot tracks project->server in `eglot--servers-by-project`.
Presently, if we xref to a file, we try to adopt it's server,
file->server is tracked in `eglot--servers-by-xreffed-file`.

But there's an issue: when we open another file of that project -
server isn't adopted.

This commit makes it so we adopt server if we've xreffed to a project,
if no project found - then we fallback to the present adoption logic of
file->server.